Update Bayonetta's Contrasty pack and fixes for previous two commits
Tried updating Contrasty to V5. I used the Default section first to fill in the values of a blank Default preset at first, but then I scrapped that since I thought that we would need to add a User-Customized preset too with all of the default values or else you're expecting users to copy-paste the default values and tweak them from there (if they see that values are being inherited).

Also forgot to change the Bayonetta and Devils Third resolution pack's version number.