Update Wind Waker HD to V5
Previous pack was weirdly ported so I redid all of the packs. Bloom is now also separated to the Mods section since it's clearly not a graphical option. I did add gamepad resolution to it. Also, the LOD Bias pack had 0.5 as one of it's presets, which just doesn't do anything since it either rounded to 0 or just flat-out didn't do anything since it's not really possible to go up half a texture level.